Cleaned up derive macro implementation
