+++ datasize_macro/Cargo.toml
@@ -11,3 +11,4 @@ proc-macro = true
 [dependencies]
 syn = { version = "1.0.103", features = ["full"] }
 quote = "1.0.21"
+proc-macro2 = "1.0.47"
